Added Breadcrumb Navigation plus mixed changes

+++ templates/portal/onappcdn/cdn_resources.tpl
@@ -41,7 +41,7 @@
         {foreach item=resource key=resource_id from=$resources}
         <tr>
             <td>
-               <a href="{$smarty.const.ONAPPCDN_FILE_NAME}?page=resources&id={$id}&action=details&resource_id={$resource_id}">
+               <a href="{$smarty.const.ONAPPCDN_FILE_NAME}?page=details&id={$id}&resource_id={$resource_id}">
                    {$resource._cdn_hostname}
                </a>
             </td>
@@ -50,7 +50,7 @@
             <td>{$resource._last_24h_cost}</td>
             <td>
                 <a href="{$smarty.const.ONAPPCDN_FILE_NAME}?page=resources&id={$id}&action=edit&resource_id={$resource_id}">{$_LANG.onappcdnedit}</a> &nbsp
-                <a href="{$smarty.const.ONAPPCDN_FILE_NAME}?page=resources&id={$id}&action=delete&resource_id={$resource_id}">{$_LANG.onappcdndelete}</a>
+                <a onclick="if( confirm('{$_LANG.onappcdnareyousureyouwantdelete}') ) return true; else return false" href="{$smarty.const.ONAPPCDN_FILE_NAME}?page=resources&id={$id}&action=delete&resource_id={$resource_id}">{$_LANG.onappcdndelete}</a>
             </td>
         </tr>
         {/foreach}
